+++ lib/index.js
@@ -1,2 +1,2 @@
-export { RootStyle } from './system';
+export { setThemeVars, RootStyle, variantColors, sizes, } from './system';
