+++ lib/system.js
@@ -126,3 +126,44 @@ export const spacePropTypes = {
   mx: PropTypes.number,
   my: PropTypes.number,
 };
+
+export const darkText = (background) => ({ hasLightText: false, background })
+export const lightText = (background) => ({ hasLightText: true, background })
+
+export const colorSystem = (options) => {
+  const fixedTextColors = options.dark ? {
+    darkText: options.background,
+    darkSecondary: options.secondaryBackground,
+    lightText: options.primary,
+    lightSecondary: options.secondary,
+  } : {
+    darkText: options.primary,
+    darkSecondary: options.secondary,
+    lightText: options.background,
+    lightSecondary: options.secondaryBackground,
+  }
+  
+  const withContext = ({ hasLightText, background }) => ({
+    text: hasLightText ? fixedTextColors.lightText : fixedTextColors.darkText,
+    secondary: hasLightText ? fixedTextColors.lightSecondary : fixedTextColors.darkSecondary,
+    background: background,
+    hoverFilter: hasLightText ? 'brightness(110%)' : 'brightness(90%)',
+  })
+  
+  return {
+    primary: options.primary,
+    secondary: options.secondary,
+    placeholder: options.placeholder,
+    border: options.border,
+    secondaryBackground: options.secondaryBackground,
+    background: options.background,
+    tertiary: withContext(options.tertiary),
+    private: withContext(options.private),
+    cta: withContext(options.cta),
+    error: withContext(options.error),
+    warning: withContext(options.warning),
+    info: withContext(options.info),
+    success: withContext(options.success),
+    selected: withContext(options.selected),
+  }
+}
