+++ lib/themes.js
@@ -1,3 +1,5 @@
+import { darkText, lightText, colorSystem } from './system'
+
 const baseTheme = {
   space: ['0', '8px', '16px', '32px', '64px', '128px'],
   fontSizes: {
@@ -26,47 +28,6 @@ const baseTheme = {
   rounded: '5px',
   popShadow: '0px 2px 5px 0px rgba(0,0,0,0.27), 0px 1px 1px 0px rgba(0,0,0,0.15)',
 };
-
-const darkText = (background) => ({ hasLightText: false, background })
-const lightText = (background) => ({ hasLightText: true, background })
-
-const colorSystem = (options) => {
-  const fixedTextColors = options.dark ? {
-    darkText: options.background,
-    darkSecondary: options.secondaryBackground,
-    lightText: options.primary,
-    lightSecondary: options.secondary,
-  } : {
-    darkText: options.primary,
-    darkSecondary: options.secondary,
-    lightText: options.background,
-    lightSecondary: options.secondaryBackground,
-  }
-  
-  const withContext = ({ hasLightText, background }) => ({
-    text: hasLightText ? fixedTextColors.lightText : fixedTextColors.darkText,
-    secondary: hasLightText ? fixedTextColors.lightSecondary : fixedTextColors.darkSecondary,
-    background: background,
-    hoverFilter: hasLightText ? 'brightness(110%)' : 'brightness(90%)',
-  })
-  
-  return {
-    primary: options.primary,
-    secondary: options.secondary,
-    placeholder: options.placeholder,
-    border: options.border,
-    secondaryBackground: options.secondaryBackground,
-    background: options.background,
-    tertiary: withContext(options.tertiary),
-    private: withContext(options.private),
-    cta: withContext(options.cta),
-    error: withContext(options.error),
-    warning: withContext(options.warning),
-    info: withContext(options.info),
-    success: withContext(options.success),
-    selected: withContext(options.selected),
-  }
-}
 
 export const lightTheme = {
   ...baseTheme,
